Bug 941004 - Refactor add browser bookmark into an app object
